[frontendActions][pull] Extract pull backgroundable to separate class
